Use INT id per brief + updated DB dump

+++ db/aeroparker.sql
@@ -25,7 +25,7 @@ DROP TABLE IF EXISTS `customers`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `customers` (
-  `id` bigint NOT NULL AUTO_INCREMENT,
+  `id` int NOT NULL AUTO_INCREMENT,
   `registered` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `email_address` varchar(255) NOT NULL,
   `title` varchar(5) NOT NULL,
@@ -61,4 +61,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2025-08-15 13:45:18
+-- Dump completed on 2025-08-15 14:29:11
